Prefer UTF8 over ASCII or 1252 when valid.
Postprocess MLang's detection:  If the analyzed chunk is ASCII or 1252,
and has at least one byte with the high bit set, and conforms to the
UTF8 encoding scheme, then assume it's UTF8.

For example, this chooses UTF8 instead of 1252 when a file is UTF8 and
contains mostly ASCII except for one emoji.  MLang isn't confident
enough to call that UTF8, but for this file viewer that's a strong
indicator (the user can always override the encoding if desired).